HADOOP-16665. Move tests into existing suite; delete new one
